+++ client/src/ts/features/notifications/state/actions.ts
@@ -1,0 +1,5 @@
+import { NotificationsSlice } from "./state"
+
+export const NotificationsActions = {
+    ...NotificationsSlice.actions,
+}