+++ client/src/ts/features/notifications/state/index.ts
@@ -1,0 +1,5 @@
+export { NotificationsActions } from "./actions"
+export { NotificationsSelectors } from "./selectors"
+
+export { NotificationsSlice } from "./state"
+export type { NotificationsState } from "./state"