+++ client/src/ts/features/notifications/state/selectors.ts
@@ -1,0 +1,23 @@
+import type { RootState } from "@app/state"
+
+const _selectSlice = (rootState: RootState) => rootState?.Notifications
+
+const selectNotifications = (rootState: RootState) => _selectSlice(rootState)?.notifications ?? []
+
+// const selectActiveAreaNavData = (rootState: RootState) => {
+//     const activeAreaName = selectActiveAreaName(rootState);
+//     if (isBlank(activeAreaName)) {
+//         return null;
+//     }
+//     if (activeAreaName === "None") {
+//         return null;
+//     }
+//     return NAV_ITEMS[activeAreaName] ?? null;
+// };
+
+
+// =============================================================================
+
+export const NotificationsSelectors = {
+    notifications: selectNotifications,
+}